requirements.txt
numpy
pandas
pyyaml
prometheus-client
